+++ src/bus/four_wire.rs
@@ -1,8 +1,7 @@
 #![allow(clippy::inconsistent_digit_grouping, clippy::unusual_byte_groupings)]
 
 use core::fmt;
-use embedded_hal::blocking::spi::{Transfer, Write};
-use embedded_hal::digital::v2::OutputPin;
+use embedded_hal::spi::{ErrorType, Operation, SpiDevice};
 
 use crate::bus::Bus;
 
@@ -11,107 +10,53 @@ const WRITE_MODE_MASK: u8 = 0b00000_1_00;
 // TODO This name is not ideal, should be renamed to VDM
 #[derive(Debug)]
 #[cfg_attr(feature = "defmt", derive(defmt::Format))]
-pub struct FourWire<Spi: Transfer<u8> + Write<u8>, ChipSelect: OutputPin> {
-    cs: ChipSelect,
-    spi: Spi,
+pub struct FourWire<SPI> {
+    spi: SPI,
 }
 
-impl<Spi: Transfer<u8> + Write<u8>, ChipSelect: OutputPin> FourWire<Spi, ChipSelect> {
-    pub fn new(spi: Spi, cs: ChipSelect) -> Self {
-        Self { cs, spi }
+impl<SPI> FourWire<SPI> {
+    pub fn new(spi: SPI) -> Self {
+        Self { spi }
     }
 
-    pub fn release(self) -> (Spi, ChipSelect) {
-        (self.spi, self.cs)
+    pub fn release(self) -> SPI {
+        self.spi
     }
 }
 
-impl<Spi: Transfer<u8> + Write<u8>, ChipSelect: OutputPin> Bus for FourWire<Spi, ChipSelect> {
-    type Error =
-        FourWireError<<Spi as Transfer<u8>>::Error, <Spi as Write<u8>>::Error, ChipSelect::Error>;
+impl<SPI: SpiDevice> Bus for FourWire<SPI> {
+    type Error = <SPI as ErrorType>::Error;
 
-    fn read_frame(&mut self, block: u8, address: u16, data: &mut [u8]) -> Result<(), Self::Error> {
+    fn read_frame(&mut self, block: u8, address: u16, data: &mut [u8]) -> Result<(), SPI::Error> {
         let address_phase = address.to_be_bytes();
         let control_phase = block << 3;
-        let data_phase = data;
 
-        // set Chip select to Low, i.e. prepare to receive data
-        self.cs.set_low().map_err(FourWireError::ChipSelectError)?;
-        let result = (|| {
-            self.spi
-                .write(&address_phase)
-                .and_then(|_| self.spi.write(&[control_phase]))
-                .map_err(FourWireError::WriteError)?;
-            self.spi
-                .transfer(data_phase)
-                .map_err(FourWireError::TransferError)?;
-            Ok(())
-        })();
+        self.spi.transaction(&mut [
+            Operation::Write(&address_phase),
+            Operation::Write(&[control_phase]),
+            Operation::TransferInPlace(data),
+        ])?;
 
-        // set Chip select to High, i.e. we've finished listening
-        self.cs.set_high().map_err(FourWireError::ChipSelectError)?;
-
-        // then return the result of the transmission
-        result
+        Ok(())
     }
 
-    fn write_frame(&mut self, block: u8, address: u16, data: &[u8]) -> Result<(), Self::Error> {
-        let address_phase = address.to_be_bytes();
+    fn write_frame(&mut self, block: u8, address: u16, data: &[u8]) -> Result<(), SPI::Error> {
         let control_phase = block << 3 | WRITE_MODE_MASK;
-        let data_phase = data;
 
-        // set Chip select to Low, i.e. prepare to transmit
-        self.cs.set_low().map_err(FourWireError::ChipSelectError)?;
-        let result = self
-            .spi
-            .write(&address_phase)
-            .and_then(|_| self.spi.write(&[control_phase]))
-            .and_then(|_| self.spi.write(data_phase))
-            .map_err(FourWireError::WriteError);
+        let address_phase = address.to_be_bytes();
+        self.spi.transaction(&mut [
+            Operation::Write(&address_phase),
+            Operation::Write(&[control_phase]),
+            Operation::Write(data),
+        ])?;
 
-        // set Chip select to High, i.e. we've finished transmitting
-        self.cs.set_high().map_err(FourWireError::ChipSelectError)?;
-
-        // then return the result of the transmission
-        result
+        Ok(())
     }
 }
-
-// Must use map_err, ambiguity prevents From from being implemented
-#[repr(u8)]
-#[derive(Clone)]
-pub enum FourWireError<TransferError, WriteError, ChipSelectError> {
-    TransferError(TransferError),
-    WriteError(WriteError),
-    ChipSelectError(ChipSelectError),
-}
-
-impl<TransferError, WriteError, ChipSelectError> fmt::Debug
-    for FourWireError<TransferError, WriteError, ChipSelectError>
-{
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(
-            f,
-            "FourWireError::{}",
-            match self {
-                Self::TransferError(_) => "TransferError",
-                Self::WriteError(_) => "WriteError",
-                Self::ChipSelectError(_) => "ChipSelectError",
-            }
-        )
-    }
-}
-
-// TODO Improved error rendering could be done with specialization.
-// https://github.com/rust-lang/rust/issues/31844
 
 #[cfg(test)]
 mod test {
-    use embedded_hal::digital::v2::OutputPin;
-    use embedded_hal_mock::{
-        pin::{Mock as PinMock, State as PinState, Transaction as PinTransaction},
-        spi::{Mock as SpiMock, Transaction as SpiTransaction},
-    };
+    use embedded_hal_mock::eh1::spi::{Mock as SpiMock, Transaction as SpiTransaction};
 
     use crate::{
         bus::{four_wire::WRITE_MODE_MASK, Bus},
@@ -122,30 +67,20 @@ mod test {
 
     #[test]
     fn test_read_frame() {
-        let mut cs_pin = PinMock::new(&[
-            // we begin with pin HIGH
-            PinTransaction::set(PinState::High),
-            // When reading
-            PinTransaction::set(PinState::Low),
-            // When finished reading
-            PinTransaction::set(PinState::High),
-        ]);
-
-        // initiate the pin to high.
-        cs_pin.set_high().expect("Should set pin to high");
-
         let mut actual_version = [0_u8; 1];
         let mut expected_version = 5;
 
         let expectations = [
-            SpiTransaction::write(register::common::VERSION.to_be_bytes().to_vec()),
-            SpiTransaction::write(vec![register::COMMON << 3]),
-            SpiTransaction::transfer(actual_version.to_vec(), vec![expected_version]),
+            SpiTransaction::transaction_start(),
+            SpiTransaction::write_vec(register::common::VERSION.to_be_bytes().to_vec()),
+            SpiTransaction::write(register::COMMON << 3),
+            SpiTransaction::transfer_in_place(actual_version.to_vec(), vec![expected_version]),
+            SpiTransaction::transaction_end(),
         ];
 
         let mock_spi = SpiMock::new(&expectations);
 
-        let mut four_wire = FourWire::new(mock_spi, cs_pin);
+        let mut four_wire = FourWire::new(mock_spi);
 
         four_wire.read_frame(
             register::COMMON,
@@ -153,41 +88,35 @@ mod test {
             &mut actual_version,
         );
 
+        four_wire.release().done();
+
         assert_eq!(expected_version, actual_version[0]);
     }
 
     #[test]
     fn test_write_frame() {
-        let mut cs_pin = PinMock::new(&[
-            // we begin with pin HIGH
-            PinTransaction::set(PinState::High),
-            // When reading
-            PinTransaction::set(PinState::Low),
-            // When finished reading
-            PinTransaction::set(PinState::High),
-        ]);
-
-        // initiate the pin to high.
-        cs_pin.set_high().expect("Should set pin to high");
-
         let socket_0_reg = 0x01_u8;
         let socket_1_reg = 0x05_u8;
         let source_port = 49849_u16;
 
         let expectations = [
-            SpiTransaction::write(register::socketn::SOURCE_PORT.to_be_bytes().to_vec()),
-            SpiTransaction::write(vec![socket_1_reg << 3 | WRITE_MODE_MASK]),
-            SpiTransaction::write(source_port.to_be_bytes().to_vec()),
+            SpiTransaction::transaction_start(),
+            SpiTransaction::write_vec(register::socketn::SOURCE_PORT.to_be_bytes().to_vec()),
+            SpiTransaction::write(socket_1_reg << 3 | WRITE_MODE_MASK),
+            SpiTransaction::write_vec(source_port.to_be_bytes().to_vec()),
+            SpiTransaction::transaction_end(),
         ];
 
         let mock_spi = SpiMock::new(&expectations);
 
-        let mut four_wire = FourWire::new(mock_spi, cs_pin);
+        let mut four_wire = FourWire::new(mock_spi);
 
         four_wire.write_frame(
             socket_1_reg,
             register::socketn::SOURCE_PORT,
             &source_port.to_be_bytes(),
         );
+
+        four_wire.release().done();
     }
 }